wasmtime_internal_unwinder

Function compute_throw_action

pub unsafe fn compute_throw_action<'a, F: Fn(usize) -> Option<(usize, ExceptionTable<'a>)>>(
    unwind: &dyn Unwind,
    module_lookup: F,
    exit_pc: usize,
    exit_frame: usize,
    entry_frame: usize,
    tag: u32,
) -> ThrowAction

Implementation of stack-walking to find a handler.

This function searches for a handler in the given range of stack frames, starting from the throw stub and up to a specified entry frame.

Safety

The safety of this function is the same as crate::visit_frames where the values passed in configuring the frame pointer walk must be correct and Wasm-defined for this to not have UB.