wasmtime_internal_unwinder

Function visit_frames

pub unsafe fn visit_frames<R>(
    unwind: &dyn Unwind,
    pc: usize,
    fp: usize,
    trampoline_fp: usize,
    f: impl FnMut(Frame) -> ControlFlow<R>,
) -> ControlFlow<R>

Walk through a contiguous sequence of Wasm frames starting with the frame at the given PC and FP and ending at trampoline_fp. This FP should correspond to that of a trampoline that was used to enter the Wasm code.

We require that the initial PC, FP, and trampoline_fp values are non-null (non-zero).

Safety

This function is not safe as unwind, pc, fp, and trampoline_fp must all be “correct” in that if they’re wrong or mistakenly have the wrong value then this method may segfault. These values must point to valid Wasmtime compiled code which respects the frame pointers that Wasmtime currently requires.